feat: refactored unit tests
