+++ dataspace-protocol-lib-mongodb/src/main/java/org/factoryx/library/connector/embedded/model/MongoTransferRecord.java
