+++ dataspace-protocol-lib-mongodb/src/main/java/org/factoryx/library/connector/embedded/mongodb/MongoNegotiationRecordFactory.java
