+++ dataspace-protocol-lib-sql/src/main/java/org/factoryx/library/connector/embedded/jpa/JpaNegotiationRecordFactory.java
