--- dsp-mongodb/src/main/java/org/factoryx/library/connector/embedded/repository/MongoTransferRecordRepository.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2025. Fraunhofer-Gesellschaft zur Foerderung der angewandten Forschung e.V. (represented by Fraunhofer ISST)
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.factoryx.library.connector.embedded.repository;
-
-import org.factoryx.library.connector.embedded.model.MongoTransferRecord;
-import org.factoryx.library.connector.embedded.provider.repository.TransferRecordRepository;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface MongoTransferRecordRepository extends MongoRepository<MongoTransferRecord, UUID>, TransferRecordRepository {
-    @Override
-    Optional<MongoTransferRecord> findById(UUID id);
-}
