+++ settings.gradle.kts
@@ -1,6 +1,4 @@
 rootProject.name = "dataspace-protocol-lib"
-include("dsp-mongodb")
-include("dsp-jpa")
 include("dsp-lib")
 include("dataspace-protocol-lib-sql")
 include("dataspace-protocol-lib-mongodb")
